Fix library includes
If I add all these libs explicitly, $brew_prefix/include is not passed to gcc.
This fixes the conflict with the original libevent found at that path.
The cmake patch at the bottom are needed for hhvm so we can pass the binutils include path.
The other patch fixes the freetype include, this is the official way of including
freetype. These patches will be submitted to hhvm upstream so they should be temporary.

With these changes I can build hhvm on my 10.9 machine with a custom brew prefix.